move parity task to lemur/common/monitoring.py
Splits pure-observability Celery tasks (read DB, emit metrics, no state
change) into their own module. lemur/common/celery.py was mixing
business-logic tasks (cert rotation, source sync) with monitoring
tasks; this PR starts the separation by putting the new
check_source_destination_parity task in lemur/common/monitoring.py.

Existing monitoring-only tasks
(report_celery_last_success_metrics, report_*_task signal handlers)
stay in celery.py for now -- they're referenced by the Celery beat
schedule in k8s-resources by full module path, so moving them is a
coordinated cross-repo change worth doing as a separate cleanup PR.
